R2cum,pcaRes-method {pcaMethods}	R Documentation

Cumulative R2 is the total ratio of variance that is being explained by the model

Description

Cumulative R2 is the total ratio of variance that is being explained by the model

Usage

## S4 method for signature 'pcaRes'
R2cum(object, ...)

Arguments

object
a pcaRes model

...
Not used

Value

Get the cumulative R2

Author(s)

Henning Redestig

[Package pcaMethods version 1.76.0 Index]