asExprSet {pcaMethods}	R Documentation

Convert pcaRes object to an expression set

Description

This function can be used to conveniently replace the expression matrix in an ExpressionSet with the completed data from a pcaRes object.

Usage

asExprSet(object, exprSet)

Arguments

object
pcaRes – The object containing the completed data.

exprSet
ExpressionSet – The object passed on to pca for missing value estimation.

Details

This is not a standard as function as pcaRes object alone not can be converted to an ExpressionSet (the pcaRes object does not hold any phenoData for example).

Value

An object without missing values of class ExpressionSet.

Author(s)

Wolfram Stacklies
CAS-MPG Partner Institute for Computational Biology, Shanghai, China

[Package pcaMethods version 1.76.0 Index]