biplot-methods {pcaMethods}	R Documentation

Plot a overlaid scores and loadings plot

Description

Visualize two-components simultaneously

Usage

## S3 method for class 'pcaRes'
biplot(x, choices = 1:2, scale = 1,
  pc.biplot = FALSE, ...)

## S4 method for signature 'pcaRes'
biplot(x, choices = 1:2, scale = 1,
  pc.biplot = FALSE, ...)

Arguments

x
a pcaRes object

choices
which two pcs to plot

scale
The variables are scaled by lambda^scale and the observations are scaled by lambda ^ (1-scale) where lambda are the singular values as computed by princomp. Normally 0 <= scale <= 1, and a warning will be issued if the specified 'scale' is outside this range.

pc.biplot
If true, use what Gabriel (1971) refers to as a "principal component biplot", with lambda = 1 and observations scaled up by sqrt(n) and variables scaled down by sqrt(n). Then the inner products between variables approximate covariances and distances between observations approximate Mahalanobis distance.

...
optional arguments to be passed to biplot.default.

Details

This is a method for the generic function 'biplot'. There is considerable confusion over the precise definitions: those of the original paper, Gabriel (1971), are followed here. Gabriel and Odoroff (1990) use the same definitions, but their plots actually correspond to pc.biplot = TRUE.

Value

a plot is produced on the current graphics device.

Author(s)

Kevin Wright, Adapted from biplot.prcomp

See Also

prcomp, pca, princomp

Examples

data(iris)
pcIr <- pca(iris[,1:4])
biplot(pcIr)

[Package pcaMethods version 1.78.0 Index]