completeObs,nniRes-method {pcaMethods}	R Documentation

Get the original data with missing values replaced with predicted values.

Description

Get the original data with missing values replaced with predicted values.

Usage

completeObs(object, ...)

Arguments

object
object to fetch complete data from

...
Not used

Value

Completed data (matrix)

Author(s)

Henning Redestig

[Package pcaMethods version 1.78.0 Index]