deletediagonals {pcaMethods}	R Documentation

Delete diagonals

Description

Replace a diagonal of elements of a matrix with NA

Usage

deletediagonals(x, diagonals = 1)

Arguments

x
The matrix

diagonals
The diagonal to be replaced, i.e. the first, second and so on when looking at the fat version of the matrix (transposed or not) counting from the bottom. Can be a vector to delete more than one diagonal.

Details

Used for creating artifical missing values in matrices without causing any full row or column to be completely missing

Value

The original matrix with some values missing

Author(s)

Henning Redestig

[Package pcaMethods version 1.78.0 Index]