fitted-methods {pcaMethods}	R Documentation

Extract fitted values from PCA.

Description

Fitted values of a PCA model

Usage

## S3 method for class 'pcaRes'
fitted(object, data = NULL, nPcs = nP(object),
  pre = TRUE, post = TRUE, ...)

## S4 method for signature 'pcaRes'
fitted(object, data = NULL, nPcs = nP(object),
  pre = TRUE, post = TRUE, ...)

Arguments

object
the pcaRes object of interest.

data
For standard PCA methods this can safely be left null to get scores x loadings but if set, then the scores are obtained by projecting provided data onto the loadings. If data contains missing values the result will be all NA. Non-linear PCA is an exception, here if data is NULL then data is set to the completeObs and propaged through the network.

nPcs
The number of PC's to consider

pre
pre-process data based on the pre-processing chosen for the PCA model

post
unpre-process the final data (add the center back etc to get the final estimate)

...
Not used

Details

This function extracts the fitted values from a pcaResobject. For PCA methods like SVD, Nipals, PPCA etc this is basically just the scores multipled by the loadings and adjusted for pre-processing. for non-linear PCA the original data is propagated through the network to obtain the approximated data.

Value

A matrix representing the fitted data

Author(s)

Henning Redestig

Examples

pc <- pca(iris[,1:4], nPcs=4, center=TRUE, scale="uv")
sum( (fitted(pc) - iris[,1:4])^2 )

[Package pcaMethods version 1.76.0 Index]