listPcaMethods {pcaMethods}	R Documentation

List PCA methods

Description

Vector with current valid PCA methods

Usage

listPcaMethods(which = c("all", "linear", "nonlinear"))

Arguments

which
the type of methods to get. E.g. only get the PCA methods based on the classical model where the fitted data is a direct multiplication of scores and loadings.

Value

A character vector with the current methods for doing PCA

Author(s)

Henning Redestig

[Package pcaMethods version 1.78.0 Index]