metaboliteDataComplete {pcaMethods}	R Documentation

A complete metabolite data set from an Arabidopsis coldstress experiment

Description

A complete subset from a larger metabolite data set. This is the original, complete data set and can be used to compare estimation results created with the also provided incomplete data (called metaboliteData). The data was created during an in house Arabidopsis coldstress experiment.

Details

A matrix containing 154 observations (rows) and 52 metabolites (columns).

Author(s)

Wolfram Stacklies

References

Matthias Scholz, Fatma Kaplan, Charles L. Guy, Joachim Kopka and Joachim Selbig. - Non-linear PCA: a missing data approach.Bioinformatics 2005 21(20):3887-3895

See Also

metaboliteData

[Package pcaMethods version 1.78.0 Index]