nVar,pcaRes-method {pcaMethods}	R Documentation

Get the number of variables used to build the PCA model.

Description

Get the number of variables used to build the PCA model.

Usage

nVar(object, ...)

Arguments

object
pcaRes object

...
Not used

Value

Number of variables

Author(s)

Henning Redestig

[Package pcaMethods version 1.76.0 Index]