nniRes {pcaMethods}	R Documentation

Class for representing a nearest neighbour imputation result

Description

This is a class representation of nearest neighbour imputation (nni) result

Details

Creating Objects
new("nniRes", completeObs=[the estimated complete observations], k=[cluster size], nObs=[amount of observations], nVar=[amount of variables], centered=[was the data centered befor running LLSimpute], center=[original means], method=[method used to perform clustering], missing=[amount of NAs])

Slots

completeObs

"matrix", the estimated complete observations

nObs

"numeric", amount of observations

nVar

"numeric", amount of variables

correlation

"character", the correlation method used (pearson, kendall or spearman)

centered

"logical", data was centered or not

center

"numeric", the original variable centers

k

"numeric", cluster size

method

"character", the method used to perform the clustering

missing

"numeric", the total amount of missing values in original data

Methods

print

Print function

Author(s)

Wolfram Stacklies

[Package pcaMethods version 1.76.0 Index]