orth {pcaMethods}	R Documentation

Calculate an orthonormal basis

Description

ONB = orth(mat) is an orthonormal basis for the range of matrix mat. That is, ONB' * ONB = I, the columns of ONB span the same space as the columns of mat, and the number of columns of ONB is the rank of mat.

Usage

orth(mat, skipInac = FALSE)

Arguments

mat
matrix to calculate orthonormal base

skipInac
do not include components with precision below .Machine$double.eps if TRUE

Value

orthonormal basis for the range of matrix

Author(s)

Wolfram Stacklies

[Package pcaMethods version 1.78.0 Index]