pcaMethods {pcaMethods}	R Documentation

pcaMethods

Description

Principal Component Analysis in R

Details

Package: 	 pcaMethods

Type: 	 Package

Developed since: 	 2006

License: 	 GPL (>=3)

LazyLoad: 	 yes

Provides Bayesian PCA, Probabilistic PCA, Nipals PCA, Inverse Non-Linear PCA and the conventional SVD PCA. A cluster based method for missing value estimation is included for comparison. BPCA, PPCA and NipalsPCA may be used to perform PCA on incomplete data as well as for accurate missing value estimation. A set of methods for printing and plotting the results is also provided. All PCA methods make use of the same data structure (pcaRes) to provide a unique interface to the PCA results. Developed at the Max-Planck Institute for Molecular Plant Physiology, Golm, Germany, RIKEN Plant Science Center Yokohama, Japan, and CAS-MPG Partner Institute for Computational Biology (PICB) Shanghai, P.R. China

Author(s)

Wolfram Stacklies, Henning Redestig

[Package pcaMethods version 1.78.0 Index]