pcaMethods-deprecated {pcaMethods}	R Documentation

Deprecated methods for pcaMethods

Description

plotR2

Lack of relevance for this plot and the fact that it can not show cross-validation based diagnostics in the same plot makes it redundant with the introduction of a dedicated plot function for pcaRes. The new plot only shows R2cum but the result is pretty much the same.

Author(s)

Henning Redestig

[Package pcaMethods version 1.76.0 Index]