pcaRes {pcaMethods}	R Documentation

Class for representing a PCA result

Description

This is a class representation of a PCA result

Details

Creating Objects
new("pcaRes", scores=[the scores], loadings=[the loadings], nPcs=[amount of PCs], R2cum=[cumulative R2], nObs=[amount of observations], nVar=[amount of variables], R2=[R2 for each individual PC], sDev=[stdev for each individual PC], centered=[was data centered], center=[original means], varLimit=[what variance limit was exceeded], method=[method used to calculate PCA], missing=[amount of NAs], completeObs=[estimated complete observations])

Slots

scores

"matrix", the calculated scores

loadings

"matrix", the calculated loadings

R2cum

"numeric", the cumulative R2 values

sDev

"numeric", the individual standard deviations of the score vectors

R2

"numeric", the individual R2 values

cvstat

"numeric", cross-validation statistics

nObs

"numeric", number of observations

nVar

"numeric", number of variables

centered

"logical", data was centered or not

center

"numeric", the original variable centers

scaled

"logical", data was scaled or not

scl

"numeric", the original variable scales

varLimit

"numeric", the exceeded variance limit

nPcs,nP

"numeric", the number of calculated PCs

method

"character", the method used to perform PCA

missing

"numeric", the total amount of missing values in original data

completeObs

"matrix", the estimated complete observations

network

"nlpcaNet", the network used by non-linear PCA

Methods (not necessarily exhaustive)

print

Print function

summary

Extract information about PC relevance

screeplot

Plot a barplot of standard deviations for PCs

slplot

Make a side by side score and loadings plot

nPcs

Get the number of PCs

nObs

Get the number of observations

cvstat

Cross-validation statistics

nVar

Get the number of variables

loadings

Get the loadings

scores

Get the scores

dim

Get the dimensions (number of observations, number of features)

centered

Get a logical indicating if centering was done as part of the model

center

Get the averages of the original variables.

completeObs

Get the imputed data set

method

Get a string naming the used PCA method

sDev

Get the standard deviations of the PCs

scaled

Get a logical indicating if scaling was done as part of the model

scl

Get the scales of the original variablesb

R2cum

Get the cumulative R2

Author(s)

Henning Redestig

[Package pcaMethods version 1.76.0 Index]