plot.pcaRes {pcaMethods}	R Documentation

Plot diagnostics (screeplot)

Description

Plot the computed diagnostics of PCA model to get an idea of their importance. Note though that the standard screeplot shows the standard deviations for the PCs this method shows the R2 values which empirically shows the importance of the P's and is thus applicable for any PCA method rather than just SVD based PCA.

Usage

## S3 method for class 'pcaRes'
plot(x, y = NULL, main = deparse(substitute(object)),
  col = gray(c(0.9, 0.5)), ...)

Arguments

x
pcaRes The pcaRes object.

y
not used

main
title of the plot

col
Colors of the bars

...
further arguments to barplot

Details

If cross-validation was done for the PCA the plot will also show the CV based statistics. A common rule-of-thumb for determining the optimal number of PCs is the PC where the CV diagnostic is at its maximum but not very far from R^2.

Value

None, used for side effect.

Author(s)

Henning Redestig

See Also

screeplot

Examples

data(metaboliteData)
pc <- pca(t(metaboliteData), nPcs=5, cv="q2", scale="uv")
plot(pc)

[Package pcaMethods version 1.76.0 Index]