plotPcs {pcaMethods}	R Documentation

Plot many side by side scores XOR loadings plots

Description

A function that can be used to visualise many PCs plotted against each other

Usage

plotPcs(object, pcs = 1:nP(object), type = c("scores", "loadings"),
  sl = NULL, hotelling = 0.95, ...)

Arguments

object
pcaRes a pcaRes object

pcs
numeric which pcs to plot

type
character Either "scores" or "loadings" for scores or loadings plot respectively

sl
character Text labels to plot instead of a point, if NULL points are plotted instead of text

hotelling
numeric Significance level for the confidence ellipse. NULL means that no ellipse is drawn.

...
Further arguments to pairs on which this function is based.

Details

Uses pairs to provide side-by-side plots. Note that this function only plots scores or loadings but not both in the same plot.

Value

None, used for side effect.

Author(s)

Henning Redestig

See Also

prcomp, pca, princomp, slplot

Examples

data(iris)
pcIr <- pca(iris[,1:4], nPcs=3,  method="svd")
plotPcs(pcIr, col=as.integer(iris[,4]) + 1)

[Package pcaMethods version 1.76.0 Index]