simpleEllipse {pcaMethods}	R Documentation

Hotelling's T^2 Ellipse

Description

Get a confidence ellipse for uncorrelated bivariate data

Usage

simpleEllipse(x, y, alfa = 0.95, len = 200)

Arguments

x
first variable

y
second variable

alfa
confidence level of the circle

len
Number of points in the circle

Details

As described in 'Introduction to multi and megavariate data analysis using PCA and PLS' by Eriksson et al. This produces very similar ellipse as compared to the ellipse function the ellipse package except that this function assumes that and y are uncorrelated (which they of are if they are scores or loadings from a PCA).

Value

A matrix with X and Y coordinates for the circle

Author(s)

Henning Redestig

See Also

ellipse

[Package pcaMethods version 1.76.0 Index]