svdPca {pcaMethods}	R Documentation

Perform principal component analysis using singular value decomposition

Description

A wrapper function for prcomp to deliver the result as a pcaRes method. Supplied for compatibility with the rest of the pcaMethods package. It is not recommended to use this function directely but rather to use the pca() wrapper function.

Usage

svdPca(Matrix, nPcs = 2, varLimit = 1, verbose = interactive(), ...)

Arguments

Matrix
Pre-processed (centered and possibly scaled) numerical matrix samples in rows and variables as columns. No missing values allowed.

nPcs
Number of components that should be extracted.

varLimit
Optionally the ratio of variance that should be explained. nPcs is ignored if varLimit < 1

verbose
Verbose complaints to matrix structure

...
Only used for passing through arguments.

Value

A pcaRes object.

Author(s)

Henning Redestig

See Also

prcomp, princomp, pca

Examples

data(metaboliteDataComplete)
mat <- prep(t(metaboliteDataComplete))
pc <- svdPca(mat, nPcs=2)
## better use pca()
pc <- pca(t(metaboliteDataComplete), method="svd", nPcs=2)

[Package pcaMethods version 1.78.0 Index]