wasna,pcaRes-method {pcaMethods}	R Documentation

Get a matrix with indicating the elements that were missing in the input data. Convenient for estimating imputation performance.

Description

Get a matrix with indicating the elements that were missing in the input data. Convenient for estimating imputation performance.

Usage

wasna(object, ...)

Arguments

object
pcaRes object

...
Not used

Value

A matrix with logicals

Author(s)

Henning Redestig

Examples

data(metaboliteData)
data(metaboliteDataComplete)
result <- pca(metaboliteData, nPcs=2)
plot(completeObs(result)[wasna(result)], metaboliteDataComplete[wasna(result)])

[Package pcaMethods version 1.76.0 Index]