weightsAccount {pcaMethods}	R Documentation

Create an object that holds the weights for nlpcaNet. Holds and sets weights in using an environment object.

Description

Create an object that holds the weights for nlpcaNet. Holds and sets weights in using an environment object.

Usage

weightsAccount(w)

Arguments

w
matrix – New weights

Value

A weightsAccound with set and current functions.

Author(s)

Henning Redestig

[Package pcaMethods version 1.78.0 Index]